+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 ## 0.9.7 - 12.03.2023
-- adds `Location` Probability and OriginSystem
+- adds `Location` Probability and OriginSystem - see [#15](https://github.com/openTdataCH/ojp-js/pull/15)
 
 ## 0.9.6 - 12.03.2023
 - improve TripRequest - see [#14](https://github.com/openTdataCH/ojp-js/pull/14)